+++ resources/js/Components/ui/sidebar/SidebarMenuButton.vue
@@ -1,0 +1,52 @@
+<script setup>
+import { reactiveOmit } from "@vueuse/core";
+import { Tooltip, TooltipContent, TooltipTrigger } from "@/Components/ui/tooltip";
+import SidebarMenuButtonChild from "./SidebarMenuButtonChild.vue";
+import { useSidebar } from "./utils.js";
+
+const props = defineProps({
+  as: { type: null, required: false, default: "button" },
+  asChild: { type: Boolean, required: false },
+  variant: { type: String, required: false, default: "default" },
+  size: { type: String, required: false, default: "default" },
+  isActive: { type: Boolean, required: false },
+  tooltip: { type: [String, Object], required: false },
+  class: {
+    type: [Boolean, null, String, Object, Array],
+    required: false,
+    skipCheck: true,
+  },
+});
+
+const { isMobile, state } = useSidebar();
+
+const delegatedProps = reactiveOmit(props, "tooltip");
+
+defineOptions({
+  inheritAttrs: false,
+});
+</script>
+
+<template>
+  <SidebarMenuButtonChild v-if="!tooltip" v-bind="{ ...delegatedProps, ...$attrs }">
+    <slot />
+  </SidebarMenuButtonChild>
+
+  <Tooltip v-else>
+    <TooltipTrigger as-child>
+      <SidebarMenuButtonChild v-bind="{ ...delegatedProps, ...$attrs }">
+        <slot />
+      </SidebarMenuButtonChild>
+    </TooltipTrigger>
+    <TooltipContent
+      side="right"
+      align="center"
+      :hidden="state !== 'collapsed' || isMobile"
+    >
+      <template v-if="typeof tooltip === 'string'">
+        {{ tooltip }}
+      </template>
+      <component :is="tooltip" v-else />
+    </TooltipContent>
+  </Tooltip>
+</template>